requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
fastapi==0.124.2
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
openai==2.11.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
